10주차 > basic > 파일 컴포넌트 디렉토리 구조화, 절대경로 설정, SearchDialog LectureTable 분리
