+++ packages/assignment-10/src/basic/main.tsx
@@ -1,4 +1,4 @@
+import App from "@/basic/App";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(<App />);
