+++ packages/assignment-10/src/basic/mocks/dummyScheduleMap.ts
