+++ packages/assignment-10/vite.config.ts
@@ -1,20 +1,25 @@
-import { defineConfig as defineTestConfig, mergeConfig } from 'vitest/config';
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react-swc';
+import react from "@vitejs/plugin-react-swc";
+import { defineConfig } from "vite";
+import { defineConfig as defineTestConfig, mergeConfig } from "vitest/config";
 
 export default mergeConfig(
   defineConfig({
     plugins: [react()],
+    resolve: {
+      alias: {
+        "@": "/src",
+      },
+    },
   }),
   defineTestConfig({
     test: {
       globals: true,
-      environment: 'jsdom',
-      setupFiles: './src/setupTests.ts',
+      environment: "jsdom",
+      setupFiles: "./src/setupTests.ts",
       coverage: {
         reportsDirectory: "./.coverage",
-        reporter: ['lcov', 'json', 'json-summary']
+        reporter: ["lcov", "json", "json-summary"],
       },
     },
   })
-)
+);
